Create the solution structure for the repo
Sets up projects for testing and the engine. More projects should be
created as is needed. The initial files were removed for cleanliness.

Enables Roslyn analyzers during the build. Some of the rules in the rule
sets don't have implementations as Roslyn analyzers, but they work in
Visual Studio so ¯\_(ツ)_/¯.