Create the langauge's scanner
Implements an LL(1) parser to convert a stream of characters to tokens.

Initially numbers scanned included the leading sign, but this was later
determined to be a very bad idea and required more context to determine
if this was a unary negate or a binary subtract.

This includes a hefty amount of unit tests and a small amount of
integration tests.