Create the base of the parser for Cimbol
Implements an LL(k) recursive descent parser for the base language
features of Cimbol. This does not include things like the await keyword,
the exists and missing operators, or other features not needed for basic
evaluation of expressions.

There's also a lot of tests in this, plus the creation of a couple of
asserts to help out with testing the parser.