Create a graph for cycle checking formulas
Implements a graph with methods for checking cycles, topological
sorting, and finding minimal partial orderings.

Cycle checking is to be used for initial validation, and minimal partial
ordering are useful for determine the execution phases of asynchronous
code.

Test coverage is at 100% for the Graph class, but some tests skimp on
verification due to difficulty reasons.