Create the API for specifying a Cimbol program
Creates a new project for the public API of Cimbol withclasses for
arguments, constants, formulas, modules, and programs.

Test coverage for this new project is currently 100%, but is entirely
unit tests.

The compile methods have not yet been built because compilation itself
has not yet been implemented and has no defined API.