Rejigger the project structure
Kills off the Cimpress.Cimbol.Compiler and Cimpress.Cimbol.Runtime
projects in favor of just moving all of the code into the main project.

Eventually, this should all be reorganized better.

Removes the ExportStatementNode class in favor of just marking Formula
nodes as exported or not.

Also adds a set of retrieval methods to the ProgramNode and
ModuleDeclarationNode classes. These will be useful when building the
dependency graph.