Create an execution plan from a dependency table
Takes a constructed dependency table and builds an execution plan. An
execution plan is a collection of groups that need to run in order to
resolve asynchronou results properly (to be implented later). It does
cool things like minimize the number of groups required by merging
adjacent execution steps that lack asynchronous formulas.

Also adds tests.