Allow for asynchornous execution of Cimbol code
Also does a lot of other things. Among these is:
* Turning off linting rules around the NotSupportedException throws.
* Flagging expression nodes as asynchronous if a child is asynchronous.
* Splitting code generation into its own file.
* Splitting function information across multiple files.

This adds support for await nodes in the AST, but currently only parses
await nodes at the root of the tree. Eventually this should be fixed to
support await nodes throughout the tree.

Adds another benchmark for async evaluation. This one averages 7.5us
locally.