Replace compile errors with descriptive errors
Replaces a subset of the placeholder NotSupportedException throws with
descriptive exceptions. This tackles just the subset of exceptions
related to compilation and parse errors.

Also adds a formula name property to the SourceText, Scanner, and
Parser classes, plus some additional information related to the current
position when parsing an expression. This was needed to provide
position information when throwing exceptions.

Initially exception messages were implemented with a resource file,
but this workflow was not ideal and discarded.

Some exceptions related to compilation and not parsing or scanning
are given garbage tracking data because that is currently not tracked
in the syntax tree.