Replace runtime errors with descriptive errors
Replaces a subset of placeholder NotSupportedException throws with
descriptive exceptions. This tackles just the subset of exceptions
related to runtime errors.

The formula name parameter on the CimbolRuntimeException is almost
always null because we don't have a good mechanism to track that yet.
Tracking formula and position information at runtime will come later.