Replace remaining placeholder errors
Replaces the last of the placeholder NotSupportedException throws with
descriptive exceptions. This tackles the internal errors and errors
that arise from improper use of the library.

This changes the behavior of TreeWalker to allow for multiple visitors
for the same event and node type combination. This was done to remove
a class of error that really was entirely artificial.