Implement a first pass at partial evaluation.
There still is work that needs to be done in regards to partial
evaluations when a compilation error occurs. Ideally, the compilation
error would be isolated to a particular formula and that would throw,
as opposed to the entire program failing. Some compilation errors, like
errors related to cycles, should not exhibit this behavior regardless.

Similarly, there needs to be a toggle for partial evaluation. In some
instances, you may want to fail immediately. The primary use case for
this is during price evaluations where we do not want the overhead of
try/catch blocks when we do not need them.

Also adds some miscellaneous tests to improve coverage.