Rejigger error output to return all errors
Previously, error output just returned the errors for exported formulas.
This was not ideal, as unexported errors were swallowed by the program.
This has been fixed by changing the program to return an
EvaluationResult object that contains the output modules and the list of
errors separate from the output modules.

It should be noted that the approach of partial evaluation is slow.
Performance should be improved by enabling a way to disable partial
evaluation in performance-critical scenarios.

Also adds some missing comments and parameterized benchmarks to measure
formula overhead.