Change IDeclarationNode inheritance tree
Reorganizes the inheritance tree for IDeclarationNode so that
ArgumentDeclarationNode, ConstantDeclarationNode, and
ModuleDeclarationNode no longer implement this interface.

Removes the string "Declaration" from the aforementioned class names, as
well as in FormulaDeclarationNode and ImportDeclarationNode.

The only structural change is that SymbolTables in a SymbolRegistry are
indexable by anything that implements ISyntaxNode instead of
IDeclarationNode to accomodate that ModuleNode no longer implements the
latter interface.