Make symbol tables less complex
Refactors the SymbolTable class to be easier to construct and work with.

Removes the notion of parent symbol tables, as they are not used. Also
removes the backreference to a symbol table's registry.

Also modifies the ExecutionStep class to not be aware of what
SymbolTable it uses, which also ended up simplifying a lot of code.