Implement a default expression
Adds a default expression for missing paths or identifiers. In and of
itself, this is a complete feature. However, this does not entirely
address the Calculator use case where we need to first re-export values
that may or may not be null.

Adds a bunch of tests, and includes changes to the scanner, parser, and
compiler.

There were discussions about having default work on values that either
errored out or were skipped. However, because of the presence of
compilation profiles, this would lead to different evaluation results in
different compilation profiles.