Re-use existing symbols for imports
Instead of constructing new symbols for imports and re-assigning values
to those, the imported symbol is reused and the additional assignment is
removed. This allows import nodes to be ignored when compiling to
bytecode.

To accomodate this, symbols are now marked if they are a reference to
another symbol. All this does is prevent a symbol created as a reference
to another symbol from being added to a LambdaExpression more than once.