Allow imports to be added to a module's exports
Enable callers to mark certain imports as exported. This will add the
result of the import to the module's list of exports, enabling the
import to be used within other modules that import the module it resides
in.

The driver for this feature is two-fold. The first is that if you have
added a module to aggregate your arguments into a single object, you'll
have to construct a simple formula for each argument in the module. This
is error prone and could be improved. The second is that you can't apply
defaults to these arguments that are propagated in this manner.